requirements.txt
fastapi==0.120.4
uvicorn==0.38.0
google-cloud-documentai==3.7.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.1
pydantic>=2.5
pydantic-settings>=2.2
google-cloud-secret-manager 
firebase-admin==6.5.0
